# openapi.yaml
openapi: 3.0.3
info:
  title: Schemathesis Go Verification API
  version: 1.0.0
  description: |
    Verification API for testing Schemathesis behavior against various OpenAPI patterns.
    This API intentionally contains both valid endpoints and buggy endpoints.

servers:
  - url: http://localhost:8080

tags:
  - name: Health
  - name: Users
  - name: Items
  - name: Orders
  - name: Auth
  - name: Bugs

security: []

paths:
  /health:
    get:
      tags: [Health]
      operationId: getHealth
      summary: Health check
      responses:
        "200":
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: role
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UserRole"
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "422":
          $ref: "#/components/responses/ValidationError"

    post:
      tags: [Users]
      operationId: createUser
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            examples:
              valid:
                value:
                  name: Alice
                  email: alice@example.com
                  age: 30
                  role: member
      responses:
        "201":
          description: Created user
          headers:
            Location:
              description: URL of created user
              schema:
                type: string
                example: /users/1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          links:
            GetCreatedUser:
              operationId: getUser
              parameters:
                userId: "$response.body#/id"
            UpdateCreatedUser:
              operationId: updateUser
              parameters:
                userId: "$response.body#/id"
            DeleteCreatedUser:
              operationId: deleteUser
              parameters:
                userId: "$response.body#/id"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"

  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          minimum: 1

    get:
      tags: [Users]
      operationId: getUser
      summary: Get user by ID
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Users]
      operationId: updateUser
      summary: Update user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      tags: [Users]
      operationId: deleteUser
      summary: Delete user
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"

  /items:
    get:
      tags: [Items]
      operationId: listItems
      summary: List items
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [book, food, tool]
        - name: minPrice
          in: query
          required: false
          schema:
            type: number
            format: float
            minimum: 0
        - name: maxPrice
          in: query
          required: false
          schema:
            type: number
            format: float
            maximum: 10000
      responses:
        "200":
          description: Item list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Item"
        "422":
          $ref: "#/components/responses/ValidationError"

    post:
      tags: [Items]
      operationId: createItem
      summary: Create item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemRequest"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
          links:
            GetCreatedItem:
              operationId: getItem
              parameters:
                itemId: "$response.body#/id"
        "422":
          $ref: "#/components/responses/ValidationError"

  /items/{itemId}:
    parameters:
      - name: itemId
        in: path
        required: true
        schema:
          type: string
          pattern: "^item_[0-9]+$"

    get:
      tags: [Items]
      operationId: getItem
      summary: Get item by ID
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          $ref: "#/components/responses/NotFound"

  /orders:
    post:
      tags: [Orders]
      operationId: createOrder
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
          links:
            GetCreatedOrder:
              operationId: getOrder
              parameters:
                orderId: "$response.body#/id"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

  /orders/{orderId}:
    parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Orders]
      operationId: getOrder
      summary: Get order by ID
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/NotFound"

  /me:
    get:
      tags: [Auth]
      operationId: getMe
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /bugs/schema-mismatch:
    get:
      tags: [Bugs]
      operationId: getSchemaMismatchBug
      summary: Intentionally returns schema mismatched response
      responses:
        "200":
          description: Should return valid user but implementation returns invalid schema
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /bugs/status-mismatch:
    get:
      tags: [Bugs]
      operationId: getStatusMismatchBug
      summary: Intentionally returns unexpected status code
      responses:
        "200":
          description: OpenAPI says 200, implementation returns 418
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /bugs/panic-on-zero:
    post:
      tags: [Bugs]
      operationId: postPanicOnZeroBug
      summary: Intentionally panics when value is zero
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  type: integer
                  minimum: -10
                  maximum: 10
      responses:
        "200":
          description: Calculation result
          content:
            application/json:
              schema:
                type: object
                required: [result]
                properties:
                  result:
                    type: integer
        "422":
          $ref: "#/components/responses/ValidationError"

  /bugs/invalid-email:
    get:
      tags: [Bugs]
      operationId: getInvalidEmailBug
      summary: Intentionally returns invalid email format
      responses:
        "200":
          description: User with invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: test-token

  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]

    UserRole:
      type: string
      enum: [admin, member, guest]

    User:
      type: object
      required: [id, name, email, age, role, createdAt]
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 50
        email:
          type: string
          format: email
        age:
          type: integer
          minimum: 0
          maximum: 120
        role:
          $ref: "#/components/schemas/UserRole"
        createdAt:
          type: string
          format: date-time

    CreateUserRequest:
      type: object
      required: [name, email, age, role]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        email:
          type: string
          format: email
        age:
          type: integer
          minimum: 0
          maximum: 120
        role:
          $ref: "#/components/schemas/UserRole"

    UpdateUserRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        email:
          type: string
          format: email
        age:
          type: integer
          minimum: 0
          maximum: 120
        role:
          $ref: "#/components/schemas/UserRole"
      minProperties: 1

    Item:
      type: object
      required: [id, name, category, price]
      properties:
        id:
          type: string
          pattern: "^item_[0-9]+$"
        name:
          type: string
          minLength: 1
          maxLength: 80
        category:
          type: string
          enum: [book, food, tool]
        price:
          type: number
          format: float
          minimum: 0
          maximum: 10000

    CreateItemRequest:
      type: object
      required: [name, category, price]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        category:
          type: string
          enum: [book, food, tool]
        price:
          type: number
          format: float
          minimum: 0
          maximum: 10000

    OrderStatus:
      type: string
      enum: [created, paid, canceled]

    Order:
      type: object
      required: [id, userId, itemIds, quantity, status]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: integer
          minimum: 1
          maximum: 2147483647
        itemIds:
          type: array
          minItems: 1
          maxItems: 5
          items:
            type: string
            pattern: "^item_[0-9]+$"
        quantity:
          type: integer
          minimum: 1
          maximum: 10
        status:
          $ref: "#/components/schemas/OrderStatus"

    CreateOrderRequest:
      type: object
      required: [userId, itemIds, quantity]
      additionalProperties: false
      properties:
        userId:
          type: integer
          minimum: 1
          maximum: 2147483647
        itemIds:
          type: array
          minItems: 1
          maxItems: 5
          items:
            type: string
            pattern: "^item_[0-9]+$"
        quantity:
          type: integer
          minimum: 1
          maximum: 10

    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - bad_request
            - unauthorized
            - not_found
            - conflict
            - validation_error
            - internal_error
        message:
          type: string
          minLength: 1
        details:
          type: array
          items:
            type: string

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"